fixes #27 Please don't use Ctrl-K for previewing
This changes the keybindings to only be active when editing
asciidoc files, and changse them to use appropriate mappings
for mac, as well as avoiding keybindings that are used elsewhere.

I've verified that these keybindings do not conflict with those
used on my system (a mac) although I've not tested against every
possible extension, and I have not tested Windows.