Allow use of ${workspaceFolder} in attributes
This allows users to define attributes in settings.json relative to
${workspaceFolder}.
It enables for example to define
"kroki-plantuml-include" : "${workspaceFolder}/style.pu"
(a path relative to the workspaceFolder),
while keeping useWorkspaceRoot = false.

Setting asciidoc.useWorkspaceRoot = true is not flexible enough for
many use cases as this changes general behaviour a lot , which is not
always desired:
e.g. include::myFile.adoc[] changes from "relative to the document"
to "relative to workspaceFolder".