fix regexes in #quote-paragraph rule, and downstream
As it stands, the regex in #quote-paragraph will not match any block-attribute-list that contains, itself, an inline macro -- such as a URL macro or a bibliographic-citation macro. Downstream from that rule, there are weaknesses also in the #block-attribute-inner rule. The fixes I've applied here are based on earlier discussion of this whole matter at https://github.com/asciidoctor/atom-language-asciidoc/pull/197 -- where there are screen-shots and syntax-examples (too bad I was barking up the wrong tree ... Atom instead of VS Code!!).